STP-245: Replace "data_type" with "precision_aggregate".
The way the application is written, raster data can only use a 32-bit floating-point representation, so it's misleading to make it seem that other representations are possible. Aggregate data _could_ be represented by other data types, but in practice we've only ever needed 32-bit or 64-bit floats, and we've never even tested other data types. To keep things simple, I think it makes sense to restrict aggregate data to single- or double-precision floats for now. That should cover the majority of use cases.