BIG5V-460: Revise LineChart config (breaking change).
The previous design was flawed. It didn't provide a good way to show multiple lines on the same chart; the "lines" array turned out not to be very useful for this. In the new design, showing multiple lines can be accomplished by specifying a "split_dimension". That means every option of the dimension named will be rendered as a separate line/area. The "lines" property has also been replaced by "area", a single object rather than an array. It's identical to the previous LineConfig type, except that "upper" and "lower" are now required properties.